+++ lib/config/react.js
@@ -5,7 +5,7 @@ module.exports = {
 
   settings: {
     react: {
-      version: '16.8',
+      version: 'detect',
     },
   },
