+++ package.json
@@ -44,7 +44,8 @@
     "extends": [
       "plugin:self/esnext",
       "plugin:self/node",
-      "plugin:self/prettier"
+      "plugin:self/prettier",
+      "plugin:self/jest"
     ]
   },
   "jest": {
@@ -56,6 +57,7 @@
     "@babel/register": "7.5.5",
     "@babel/runtime-corejs2": "^7.4.3",
     "babel-preset-shopify": "^20.0.0",
+    "chai": "^4.2.0",
     "eslint": "^6.0.1",
     "eslint-find-rules": "^3.3.1",
     "eslint-index": "^1.5.0",
